auth

+++ html-geradorUCs/routes/uc.js
@@ -5,6 +5,22 @@ var path = require('path');
 var fs = require('fs');
 
 router.get('/edit/:id', async function (req, res, next) {
+	var userLogged = false;
+	var isAdmin = false;
+	var username = "";
+
+	if (req.cookies.token != 'undefined') {
+		const response = await axios.get(process.env.AUTH_URI + '/isLogged?token=' + req.cookies.token)
+		userLogged = response.data.isLogged;
+		isAdmin = response.data.isAdmin;
+		username = response.data.username;
+	}
+
+    if (!userLogged) {
+        res.redirect('/login')
+        return;
+    }
+
     try {
         let ucResponse = await axios.get(process.env.API_URI + '/ucs/' + req.params.id);
         let docentesResponse = await axios.get(process.env.API_URI + '/docentes');
@@ -16,13 +32,29 @@ router.get('/edit/:id', async function (req, res, next) {
             return docenteResponse.data;
         }));
         
-        res.render('editUC', { uc, docentes, title: 'Editar UC: ' + uc.titulo});
+        res.render('editUC', { uc, docentes, title: 'Editar UC: ' + uc.titulo, userLogged, isAdmin, username });
     } catch (error) {
-        res.render('error', { error: { status: 501, message: 'Erro ao consultar UC ou Docentes' }, title: 'Erro' });
+        res.render('error', { error: { status: 501, message: 'Erro ao consultar UC ou Docentes' }, title: 'Erro', userLogged, isAdmin, username});
     }
 });
 
 router.post('/edit/:id', async function (req, res, next) {
+	var userLogged = false;
+	var isAdmin = false;
+	var username = "";
+
+	if (req.cookies.token != 'undefined') {
+		const response = await axios.get(process.env.AUTH_URI + '/isLogged?token=' + req.cookies.token)
+		userLogged = response.data.isLogged;
+		isAdmin = response.data.isAdmin;
+		username = response.data.username;
+	}
+
+    if (!userLogged) {
+        res.redirect('/login')
+        return;
+    }
+
     try {
         var updatedUC = {
             _id: req.body._id,
@@ -44,15 +76,26 @@ router.post('/edit/:id', async function (req, res, next) {
         await axios.put(process.env.API_URI + '/ucs/' + req.params.id, updatedUC);
         res.redirect('/uc/' + req.params.id);
     } catch (error) {
-        res.render('error', { error: { status: 501, message: 'Erro ao editar UC' }, title: 'Erro' });
+        res.render('error', { error: { status: 501, message: 'Erro ao editar UC' }, title: 'Erro', userLogged, isAdmin, username});
     }
 });
 
 router.get('/:id', async function (req, res, next) {
+	var userLogged = false;
+	var isAdmin = false;
+	var username = "";
+
+	if (req.cookies.token != 'undefined') {
+		const response = await axios.get(process.env.AUTH_URI + '/isLogged?token=' + req.cookies.token)
+		userLogged = response.data.isLogged;
+		isAdmin = response.data.isAdmin;
+		username = response.data.username;
+	}
+
 	try {		
 		let dados = await axios.get(process.env.API_URI + '/ucs/' + req.params.id);
 		if (dados.data == null) {
-			res.render('error', { error: { status: 404, message: 'UC não encontrada' }, title: 'Erro' });
+			res.render('error', { error: { status: 404, message: 'UC não encontrada' }, title: 'Erro', userLogged, isAdmin, username});
 			return;
 		}
 
@@ -63,20 +106,34 @@ router.get('/:id', async function (req, res, next) {
 				let docenteData = await axios.get(process.env.API_URI + '/docentes/' + docentes[j]);
 				newDocentes.push(docenteData.data);
 			} catch (erro) {
-				res.render('error', { error: { status: 501, message: 'Erro ao consultar Docente' }, title: 'Erro' });
+				res.render('error', { error: { status: 501, message: 'Erro ao consultar Docente' }, title: 'Erro', userLogged, isAdmin, username});
 				return;
 			}
 		}
 		dados.data.docentes = newDocentes;
 		
-		res.render('uc', { uc: dados.data, title: 'UC: ' + dados.data.titulo});
+		res.render('uc', { uc: dados.data, title: 'UC: ' + dados.data.titulo, userLogged, isAdmin, username });
 	} catch (erro) {
-		res.render('error', { error: { status: 501, message: 'Erro ao consultar UC' }, title: 'Erro' });
+		res.render('error', { error: { status: 501, message: 'Erro ao consultar UC' }, title: 'Erro', userLogged, isAdmin, username});
 	}
 });
 
-router.post('/', function (req, res, next) {
-    // TODO AUTH
+router.post('/', async function (req, res, next) {
+    var userLogged = false;
+	var isAdmin = false;
+	var username = "";
+
+	if (req.cookies.token != 'undefined') {
+		const response = await axios.get(process.env.AUTH_URI + '/isLogged?token=' + req.cookies.token)
+		userLogged = response.data.isLogged;
+		isAdmin = response.data.isAdmin;
+		username = response.data.username;
+	}
+    
+    if (!userLogged) {
+        res.redirect('/login')
+        return;
+    }
 
 	var newUC = {
 		_id: req.body._id,
@@ -101,14 +158,30 @@ router.post('/', function (req, res, next) {
             res.redirect('/uc/' + response.data._id);
         })
         .catch((error) => {
-            res.render('error', { error: { status: 501, message: 'Erro ao adicionar UC' }, title: 'Erro' });
+            res.render('error', { error: { status: 501, message: 'Erro ao adicionar UC' }, title: 'Erro', userLogged, isAdmin, username});
         });
 });
 
 router.delete('/:id', async function (req, res, next) {
+	var userLogged = false;
+	var isAdmin = false;
+	var username = "";
+
+	if (req.cookies.token != 'undefined') {
+		const response = await axios.get(process.env.AUTH_URI + '/isLogged?token=' + req.cookies.token)
+		userLogged = response.data.isLogged;
+		isAdmin = response.data.isAdmin;
+		username = response.data.username;
+	}
+
+    if (!userLogged) {
+        res.redirect('/login')
+        return;
+    }
+
 	axios.delete(process.env.API_URI + '/ucs/' + req.params.id)
 		.then(dados => res.redirect('/'))
-		.catch(erro => res.render('error', { error: { status: 501, message: 'Erro ao eliminar UC' }, title: 'Erro' }));
+		.catch(erro => res.render('error', { error: { status: 501, message: 'Erro ao eliminar UC' }, title: 'Erro', userLogged, isAdmin, username}));
 });
 
 module.exports = router;
